Added Dart to file extensions (#20)
* Added Dart to file extensions

Was working on a flutter project and was annoyed that the dart files had no icons. I did see that the dart logo icon was in this project but it was still missing in the config. I did add the dart file type to the fileExtensions.ts which fixed the problem.

* Added dart to language id